Move incomplete question state management to app.tsx
